movielens-gpt-chatbot/requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
python-dotenv
langchain-community
faiss-cpu